Worked on responsiveness of component. And updated format in registeredComponents.js
